+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-# [Unreleased]
+## [Unreleased]
+### Added
+* Add new `Formality` options: `PreferLess` and `PreferMore`.
 ### Changed
 * Requests resulting in `503 Service Unavailable` errors are now retried.
   Attempting to download a document before translation is completed will now